backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings>=2.2.1
mcp[cli]>=1.6.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
httpx>=0.28.1
langchain-openai>=0.1.14
langchain-anthropic>=1.4.0
langchain-google-genai>=4.2.1
langchain-mistralai>=1.1.2
langgraph>=0.1.14
chromadb==0.5.3
openai>=1.35.7
python-multipart==0.0.9
pyyaml>=6.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# OpenTelemetry
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
opentelemetry-instrumentation-fastapi>=0.53b0
opentelemetry-instrumentation-httpx>=0.53b0
opentelemetry-instrumentation-sqlalchemy>=0.53b0
opentelemetry-instrumentation-asyncpg>=0.53b0
opentelemetry-instrumentation-logging>=0.53b0
